Add files via upload
The user is able to interact with the eyes by moving the curser across the screen, while doing so the eyes follows it. 
This activity was created by the MITXPRO program to demonstrate the usage of JavaScript and Dom. 
I think it would be a nice touch if the eye could blink as well, and giving it a more human like look, but that will come with time as I learn how to manipulate the language better.